--- a/README.md
+++ b/README.md
@@ -1,2 +1,5 @@
-# Digital-Twins
-Digital Twin
+# DigitalTwin_SpatialDB
+
+
+
+