--- a/src/dynamic_boundary_conditions/hirds_rainfall_data_from_db.py
+++ b/src/dynamic_boundary_conditions/hirds_rainfall_data_from_db.py
@@ -5,7 +5,6 @@
 """
 
 import logging
-import pathlib
 from typing import Optional
 
 import geopandas as gpd