--- a/src/dynamic_boundary_conditions/hirds_rainfall_data_to_db.py
+++ b/src/dynamic_boundary_conditions/hirds_rainfall_data_to_db.py
@@ -7,7 +7,6 @@
 import pandas as pd
 import geopandas as gpd
 import sqlalchemy
-import pathlib
 import logging
 from typing import List
 from src.digitaltwin import setup_environment