--- a/src/dynamic_boundary_conditions/hyetograph.py
+++ b/src/dynamic_boundary_conditions/hyetograph.py
@@ -5,8 +5,8 @@
 """
 
 import logging
-import pathlib
 import pandas as pd
+import geopandas as gpd
 import numpy as np
 from typing import Union
 from math import floor, ceil