--- a/src/dynamic_boundary_conditions/main_tide_slr.py
+++ b/src/dynamic_boundary_conditions/main_tide_slr.py
@@ -46,9 +46,8 @@ def write_nz_bbox_to_file(engine, file_name: str = "nz_bbox.geojson"):
 
 
 def get_catchment_area(
-        catchment_file: Union[str, pathlib.Path],
+        catchment_area: gpd.GeoDataFrame,
         to_crs: int = 2193) -> gpd.GeoDataFrame:
-    catchment_area = gpd.read_file(catchment_file)
     catchment_area = catchment_area.to_crs(to_crs)
     return catchment_area
 
@@ -76,13 +75,13 @@ def remove_existing_boundary_input(bg_flood_dir: pathlib.Path):
         file_path.unlink()
 
 
-def main():
+def main(selected_polygon_gdf: gpd.GeoDataFrame):
     try:
         # Connect to the database
         engine = setup_environment.get_database()
         write_nz_bbox_to_file(engine)
         # Get catchment area
-        catchment_area = get_catchment_area("selected_polygon.geojson")
+        catchment_area = get_catchment_area(selected_polygon_gdf, to_crs=2193)
         # BG-Flood Model Directory
         bg_flood_dir = config.get_env_variable("FLOOD_MODEL_DIR", cast_to=pathlib.Path)
         # Remove existing tide model input files
@@ -127,4 +126,5 @@ def main():
 
 
 if __name__ == "__main__":
-    main()
+    sample_polygon = gpd.GeoDataFrame.from_file("selected_polygon.geojson")
+    main(sample_polygon)