--- a/src/dynamic_boundary_conditions/main_tide_slr.py
+++ b/src/dynamic_boundary_conditions/main_tide_slr.py
@@ -6,7 +6,6 @@
 
 import logging
 import pathlib
-from typing import Union
 
 import sqlalchemy
 import geopandas as gpd