--- a/src/dynamic_boundary_conditions/model_input.py
+++ b/src/dynamic_boundary_conditions/model_input.py
@@ -193,8 +193,7 @@ def generate_rain_model_input(
 
 def main():
     # BG-Flood path
-    flood_model_dir = config.get_env_variable("FLOOD_MODEL_DIR")
-    bg_flood_path = pathlib.Path(flood_model_dir)
+    bg_flood_path = config.get_env_variable("FLOOD_MODEL_DIR", cast_to=pathlib.Path)
     # Catchment polygon
     catchment_gdf = gpd.GeoDataFrame.from_file("selected_polygon.geojson")
     catchment_polygon = main_rainfall.catchment_area_geometry_info(catchment_gdf)
@@ -221,7 +220,6 @@ def main():
     # Get the intersection of rainfall sites coverage areas (thiessen polygons) and the catchment area
     sites_coverage = sites_coverage_in_catchment(sites_in_catchment, catchment_polygon)
     # Write out the requested rainfall model input for BG-Flood
-    bg_flood_path = config.get_env_variable("FLOOD_MODEL_DIR", cast_to=pathlib.Path)
     generate_rain_model_input(hyetograph_data, sites_coverage, bg_flood_path, input_type=RainInputType.UNIFORM)
     generate_rain_model_input(hyetograph_data, sites_coverage, bg_flood_path, input_type=RainInputType.VARYING)