--- a/src/dynamic_boundary_conditions/thiessen_polygons.py
+++ b/src/dynamic_boundary_conditions/thiessen_polygons.py
@@ -6,7 +6,6 @@
 """
 
 import logging
-import pathlib
 
 import geopandas as gpd
 import pandas as pd