## DESCRIPTION OF PR:
* Add Tide and Sea Level Rise module `tide_slr` to the codebase
* Fix docstrings and comments in main_rainfall.py
* Update bg_flood_model.py to incorporate tide and sea-level rise model inputs

- Closes:
#95 
#97 
#62 

* Notes and To-Dos in New PR:
   * #106 The initial draft codes for the new Tide and Sea Level Rise module `tide_slr` prototype have been added. However, further enhancements are necessary to improve code quality and performance. This will involve restructuring and refactoring the `tide_slr` module codes. Additionally, comprehensive docstrings, comments, and descriptions need to be added to all `tide_slr` scripts to ensure clarity and ease of understanding.
   * #107 

## Developer Checklist
- [x] Make code change
- [ ] Update tests
  - [ ] Update / create new tests
  - [ ] Ensure these tests have the expected behaviour
  - [ ] Test locally and ensure tests are passing
- [x] Update documentation
  - [ ] Readme
  - [x] Docstrings
  - [x] Comments
  - [ ] Wiki

## Reviewer Checklist
- [ ] Check new code for code smells
- [ ] Check new tests
  - [ ] Ensure adequate coverage
  - [ ] Check for code smells within tests
- [ ] Check if documentation needs updating
  - [ ] Readme
  - [ ] Docstrings
  - [ ] Comments
  - [ ] Wiki