@LukeParky 2022-10-20
@A-Load-Of-c0de overall nice structure to the code. I think @caseyli92 and I can refactor some of it to meet our internal guidelines but that's not anything to worry about. Let us know in here when you have run out of time to work on this project.