@xandercai 2022-11-30
Reconstructed the code to follow the lint and optimize performance.
For tide_data.py: use `asyncio` and `aiohttp` request API in an async parallel way. Runtime drops from 8.22 sec to 1.32 sec.
For sea_level_rise_data.py: use `pyarrow` read CSV files, and avoid using geopandas. Runtime drops from 104.09 sec to 0.68 sec.