@caseyli92 2023-06-01
DESCRIPTION OF PR #96 (merged into this branch):
Improve code quality and performance by restructuring and refactoring tide_data.py and sea_level_rise_data.py which were previously worked on by Angus and Xander
Improve performance and runtime using the Pyarrow library directly instead of pandas with the Pyarrow engine
Change to using Pathlib to work with filesystem paths for consistency
Change to using Geopandas to calculate distance instead of using haversine library
Updated and resolved some issues/bugs in the original code
Note: A new pull request will be created to implement necessary corrections and modifications in the code as per Matt's requirements, and to incorporate any new features as needed.