@caseyli92 2023-06-01
DESCRIPTION OF PR #99 (merged into this branch):
Bring the existing Tide and Sea Level Rise codes to a stage where they can be utilized as input for the bg-flood model. This should involve implementing any required adjustments and modifications in accordance with Matt's specifications. Additionally, incorporate any new features necessary to ensure the code is thorough and current.